selenium_tests/requirements.txt
selenium==4.27.1
pytest==8.3.4
pytest-html==4.1.1
allure-pytest==2.13.5
webdriver-manager==4.0.2
pytest-xdist==3.6.1
